[Mobile] Fix QMobileDataStore TU by resetting all Core Data stack objects
Imported from Git
 Author: Eric Marchand <eric.marchand@4d.com> 1537774196 +0200
 Committer: Eric Marchand <eric.marchand@4d.com> 1537774196 +0200
 sha1: 1d92e32e7fee5f346ead89b2904103d69a26831e
 push-state: complete
 parent-changes: 10085e345a85ea32ef6419cbd9e5b9a8c6175e94=[227433]

[git-p4: depot-paths = "//depot/4eDimension/main/QMobile/QMobileDataStore/": change = 227682]

+++ Sources/CoreData/CoreDataStore.swift
@@ -168,7 +168,18 @@ import XCGLogger
     }
 
     // MARK: computed variables
-    fileprivate lazy var persistentContainer: NSPersistentContainer = {
+
+    fileprivate var _persistentContainer: NSPersistentContainer?
+    fileprivate var persistentContainer: NSPersistentContainer {
+        if let persistentContainer = _persistentContainer {
+            return persistentContainer
+        }
+        let value = setupContainer()
+        _persistentContainer = value
+        return value
+    }
+
+    private func setupContainer() -> NSPersistentContainer {
         let modelName = self.model.name()
         guard let managedObjectModel = self.model.model() else {
             return NSPersistentContainer(name: modelName)
@@ -179,9 +190,11 @@ import XCGLogger
 
         container.persistentStoreDescriptions = [description]
         logger.verbose("\(container.persistentStoreDescriptions)")
-
         return container
-    }()
+    }
+    private func unsetupContainer() {
+        _persistentContainer = nil
+    }
 
     fileprivate var persistentStoreCoordinator: NSPersistentStoreCoordinator {
         return persistentContainer.persistentStoreCoordinator
@@ -404,6 +417,7 @@ extension CoreDataStore: DataStore {
                     } catch {
                         result = .failure(DataStoreError(error))
                     }
+                    unsetupContainer()
                     Notification(name: .dataStoreDropped, object: result).post(.dataStore)
                     completionHandler?(result)
                     return
@@ -415,6 +429,8 @@ extension CoreDataStore: DataStore {
                     return
                 }
 
+                self.unsetupContainer()
+
                 // get the store url
                 guard let storeURL = store.url, storeURL.isFileURL else {
                     Notification(name: .dataStoreDropped, object: result, userInfo: ["noFile": true]).post(.dataStore)
@@ -424,10 +440,13 @@ extension CoreDataStore: DataStore {
 
                 do {
                     // self.persistentStoreCoordinator.remove(store)
-                    try self.persistentStoreCoordinator.destroyPersistentStore(at: storeURL, ofType: self.storeType.type, options: store.options)
+                    if persistentStoreCoordinator.persistentStores.first?.url != nil {
+                        print("Missing first store URL - could not destroy")
+                        try self.persistentStoreCoordinator.destroyPersistentStore(at: storeURL, ofType: self.storeType.type, options: store.options)
+                    }
+
                     // remove the files
                     try self.drop(storeURL: storeURL)
-
                 } catch {
                     result = .failure(DataStoreError(error))
                 }
